Move compare_html helper closer to call sites
