+++ tests/unit/cmk/gui/compare_html.py
