+++ tests/unit/cmk/gui/htmllib/test_foldable_container.py
@@ -5,7 +5,7 @@
 
 import pytest
 
-from tests.testlib import compare_html
+from tests.unit.cmk.gui.compare_html import compare_html
 
 from cmk.gui.htmllib.foldable_container import (
     foldable_container,
