+++ tests/unit/cmk/gui/htmllib/test_html_cls.py
@@ -7,7 +7,7 @@ import traceback
 
 import pytest
 
-from tests.testlib import compare_html
+from tests.unit.cmk.gui.compare_html import compare_html
 
 from cmk.gui.exceptions import MKUserError
 from cmk.gui.htmllib.generator import HTMLWriter
